Validar estrutura de arquivos de token
