+++ .gitignore
@@ -1,6 +1,6 @@
 # User addition
 .DS_Store
-*.ini
+config.ini
 *.json
 *.csv
