+++ scraper/badtoken.ini
@@ -1,0 +1,2 @@
+[DAFAULT]
+token = TestTokenNotValid
