+++ scraper/invalidtoken.ini
@@ -1,0 +1,3 @@
+[DEFAULT]
+token = TestTokenNotValid
+
