+++ scraper/token_manager.py
@@ -1,0 +1,31 @@
+import os
+from configparser import ConfigParser
+
+
+def retrieve_token(file='config.ini'):
+    """
+    Retrieves token from config.ini file on scraper folder
+    """
+    try:
+        config = ConfigParser()
+        config.read_file(open(str(os.getcwd())+'/scraper/'+file))
+        if ('DEFAULT' in config.keys()):
+            if ('token' in config['DEFAULT'].keys()):
+                return(config['DEFAULT']['token'])
+        return 'Token with bad structure'
+    except:
+        return False
+
+
+def update_token(new_token=None, file='config.ini'):
+    """
+    Update token from config.ini file on scraper folder
+    """
+    if new_token is not None:
+        config = ConfigParser()
+        config['DEFAULT'] = {'token': new_token}
+        with open(str(os.getcwd())+'/scraper/'+file, 'w') as configfile:
+            config.write(configfile)
+        return 'New token written successfuly.'
+    else:
+        return 'Token not updated.'
